requirements.txt
paho-mqtt==1.6.1
structlog==22.1.0
pandas==1.5.1